+++ pkg/nn/rate.go
@@ -1,10 +1,10 @@
 package nn
 
 // DefaultRate default learning rate
-const DefaultRate float32 = .3
+const DefaultRate float64 = .3
 
 // checkLearningRate
-func checkLearningRate(rate float32) float32 {
+func checkLearningRate(rate float64) float64 {
 	switch {
 	case rate < 0 || rate > 1:
 		return DefaultRate
